Auth System Updates

+++ AstroNetworkingLibrary/Serializable/AuthData.cs
@@ -6,11 +6,11 @@
     [Serializable, Obfuscation]
     public class AuthData
     {
-        public byte ClientType;
+        public int ClientType;
 
         public string Key = string.Empty;
 
-        public AuthData(byte clientType, string key = "")
+        public AuthData(int clientType, string key = "")
         {
             ClientType = clientType;
             Key = key;
